refactor(bigquery): split append builder
